Fix compatibility with fmt 8.1.1 (fixes #1653)

+++ src/output-flex.cpp
@@ -637,7 +637,7 @@ void output_flex_t::write_column(
         }
     } else {
         throw std::runtime_error{
-            "Column type {} not implemented."_format(column.type())};
+            "Column type {} not implemented."_format(static_cast<uint8_t>(column.type()))};
     }
 
     lua_pop(lua_state(), 1);
